replaced profiles with property and moved config files
